requirements.txt
openai>=1.0.0
pinecone>=3.1.0
streamlit
python-dotenv
